requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
numpy==1.26.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0
scikit-learn==1.5.0
